pmsx003: Add new aqi calculation functions
The pmsx003 sensors are quite capable of computing the AQI, but there
is no support in the driver to do this.  The hm3301 driver does support
this, but the pmsx003 is a bit different.  This brings the support over
from there, and changes it where appropriate in order to work.

The AQI computation code should probably be refactored into a standard
particulate matter parent class, but doing that is beyond the scope of
this particular pull request (and I do not have the hardware to test it
with.)